Split post 1427 oversized slices
