+++ app/src/main/java/com/pocketshell/app/tmux/TmuxSessionViewModel.kt
@@ -7504,19 +7504,7 @@ public class TmuxSessionViewModel @Inject constructor(
     private fun reconnectSourceCandidate(
         trigger: TmuxConnectTrigger?,
         sourceCandidate: String?,
-    ): String =
-        when {
-            sourceCandidate == "terminal_surface" -> "terminal_ui_failure"
-            sourceCandidate == "passive_disconnect" -> "tmux_eof_or_reader_disconnect"
-            sourceCandidate == "background_teardown" -> "lifecycle_teardown"
-            sourceCandidate == "foreground_reattach" -> "manual_or_foreground_reattach"
-            sourceCandidate == "network_observer" -> "network_replay_or_handoff"
-            trigger == TmuxConnectTrigger.LifecycleReattach -> "manual_or_foreground_reattach"
-            trigger == TmuxConnectTrigger.NetworkReconnect -> "network_replay_or_handoff"
-            trigger == TmuxConnectTrigger.Reconnect -> "manual_or_foreground_reattach"
-            trigger == TmuxConnectTrigger.AutoReconnect -> "tmux_eof_or_reader_disconnect"
-            else -> "none"
-        }
+    ): String = tmuxReconnectSourceCandidate(trigger, sourceCandidate)
 
     private fun recordAutoReconnectDecision(
         decision: String,
@@ -7627,21 +7615,14 @@ public class TmuxSessionViewModel @Inject constructor(
         event: String,
         trigger: TmuxConnectTrigger,
         detail: String = "",
-    ): String = buildString {
-        append(event)
-        append(" attempt=")
-        append(attempt)
-        append(" trigger=")
-        append(trigger.logValue)
-        append(' ')
-        append(targetLogFields(target))
-        append(" elapsedMs=")
-        append(SystemClock.elapsedRealtime() - startedAtMs)
-        if (detail.isNotBlank()) {
-            append(' ')
-            append(detail)
-        }
-    }
+    ): String = tmuxAttachMilestoneMessage(
+        attempt = attempt,
+        target = target,
+        startedAtMs = startedAtMs,
+        event = event,
+        trigger = trigger,
+        detail = detail,
+    )
 
     // Issue #1224: fed by the per-pane [TmuxClient.outputFor] tap
     // ([recordVisiblePaneOutput]), NOT by `%output` on the structural
